refactor: vo 패키지 변경, 마지막 줄이 개행으로 끝나도록 변경
