+++ src/main/java/domain/card/DeckMaker.java
@@ -1,5 +1,6 @@
 package domain.card;
 
+import domain.card.vo.Card;
 import java.util.List;
 
 public interface DeckMaker {
