+++ src/main/java/domain/card/OneRandomDeckMaker.java
@@ -1,5 +1,8 @@
 package domain.card;
 
+import domain.card.vo.Card;
+import domain.card.vo.Rank;
+import domain.card.vo.Suit;
 import java.util.Collections;
 import java.util.LinkedList;
 import java.util.List;
