+++ src/main/java/domain/state/Burst.java
@@ -14,4 +14,3 @@ public class Burst extends State {
         return true;
     }
 }
-
