+++ src/test/java/CardsCreatorTest.java
@@ -1,12 +1,12 @@
 import static org.assertj.core.api.Assertions.assertThat;
 
-import domain.card.Card;
-import util.CardsCreator;
-import domain.card.Rank;
-import domain.card.Suit;
+import domain.card.vo.Card;
+import domain.card.vo.Rank;
+import domain.card.vo.Suit;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import util.CardsCreator;
 
 class CardsCreatorTest {
